Fjarlægt skrár úr application og fært yfir í ui
