--- src/main/java/is/hi/flight_booking/application/SelectFlightController.java
@@ -1,15 +1,0 @@
-package is.hi.flight_booking.application;
-
-import javafx.fxml.Initializable;
-
-import java.net.URL;
-import java.util.ResourceBundle;
-
-public class SelectFlightController implements Initializable {
-    @Override
-    public void initialize(URL url, ResourceBundle resourceBundle) {
-        //Þessi er algerlega eftir, en planið er semsagt að breyta upphafssenunni í staðin fyrir
-        // að vera með pop-up glugga eftir pop-up glugga.
-        // Ætlunin er að senda FlightControllerinn bara á milli controllera þegar skipt er um senur.
-    }
-}
