+++ src/main/java/is/hi/flight_booking/ui/SelectFlightController.java
@@ -1,0 +1,17 @@
+package is.hi.flight_booking.ui;
+
+import javafx.fxml.Initializable;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class SelectFlightController implements Initializable {
+  @Override
+  public void initialize(URL url, ResourceBundle resourceBundle) {
+    // Þessi er algerlega eftir, en planið er semsagt að breyta upphafssenunni í
+    // staðin fyrir
+    // að vera með pop-up glugga eftir pop-up glugga.
+    // Ætlunin er að senda FlightControllerinn bara á milli controllera þegar skipt
+    // er um senur.
+  }
+}
